Corrected errors according to the comments of the mentor
