+++ src/main/java/pro/sky/animal_shelter_telegram_bot/controller/ConstantsOfControllers.java
@@ -1,0 +1,30 @@
+package pro.sky.animal_shelter_telegram_bot.controller;
+
+public class ConstantsOfControllers {
+
+    public final static String HELLO_MESSAGE_IN_MAIN_CONTROLLER = "Welcome to our pet shelter!";
+
+    public final static String HELLO_MESSAGE_OF_PET_CONTROLLER = "You can do it by information of pet:\n" +
+            "1. add pet information\n" +
+            "2. get pet information\n" +
+            "2. update pet information\n" +
+            "4. remove pet information";
+
+    public final static String HELLO_MESSAGE_OF_PET_OWNER_CONTROLLER = "You can do it by information of pet owner:\n" +
+            "1. add information about the owner of the pet\n" +
+            "2. get information about the owner of the pet\n" +
+            "2. update information about the owner of the pet\n" +
+            "4. remove information about the owner of the pet\n";
+
+    public final static String HELLO_MESSAGE_OF_REPORT_CONTROLLER = "You can do it by reports\n" +
+            "1. add new report\n" +
+            "2. find report\n" +
+            "2. update report\n" +
+            "4. remove report\n";
+
+    public final static String HELLO_MESSAGE_VOLUNTEER_CONTROLLER = "You can do it by information of volunteer:\n" +
+            "1. add information about the volunteer\n" +
+            "2. get information about the volunteer\n" +
+            "2. update information about the volunteer\n" +
+            "4. remove information about rhe volunteer\n";
+}
