+++ src/main/java/pro/sky/animal_shelter_telegram_bot/controller/PetController.java
@@ -13,16 +13,13 @@ import org.springframework.web.bind.annotation.*;
 import pro.sky.animal_shelter_telegram_bot.model.pets.Pet;
 import pro.sky.animal_shelter_telegram_bot.service.PetService;
 
+import static pro.sky.animal_shelter_telegram_bot.controller.ConstantsOfControllers.HELLO_MESSAGE_OF_PET_CONTROLLER;
+
 @RestController
 @RequestMapping("/pet")
 public class PetController {
 
     private final PetService petService;
-    private final String HELLO_MESSAGE = "You can do it by information of pet:\n" +
-            "1. add pet information\n" +
-            "2. get pet information\n" +
-            "2. update pet information\n" +
-            "4. remove pet information";
 
     public PetController(PetService petService) {
         this.petService = petService;
@@ -42,7 +39,7 @@ public class PetController {
     )
     @GetMapping
     public String helloMessage(){
-        return HELLO_MESSAGE;
+        return HELLO_MESSAGE_OF_PET_CONTROLLER;
     }
 
     @Operation(
