+++ src/main/java/pro/sky/animal_shelter_telegram_bot/controller/ReportController.java
@@ -6,39 +6,23 @@ import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.multipart.MultipartFile;
 import pro.sky.animal_shelter_telegram_bot.model.Report;
-import pro.sky.animal_shelter_telegram_bot.model.pets.PhotoOfPet;
-import pro.sky.animal_shelter_telegram_bot.service.PhotoOfPetService;
 import pro.sky.animal_shelter_telegram_bot.service.ReportService;
 
-import java.io.*;
+import static pro.sky.animal_shelter_telegram_bot.controller.ConstantsOfControllers.HELLO_MESSAGE_OF_REPORT_CONTROLLER;
 
 @RestController
 @RequestMapping("/report")
 public class ReportController {
 
     private final ReportService reportService;
-    private final PhotoOfPetService photoOfPetService;
 
-    Logger logger = LoggerFactory.getLogger(PhotoOfPetService.class);
-
-    private final String HELLO_MESSAGE = "You can do it by reports\n" +
-            "1. add new report\n" +
-            "2. find report\n" +
-            "2. update report\n" +
-            "4. remove report\n";
-
-    public ReportController(ReportService reportService, PhotoOfPetService photoOfPetService) {
+    public ReportController(ReportService reportService) {
         this.reportService = reportService;
-        this.photoOfPetService = photoOfPetService;
     }
 
     @Operation(
@@ -55,7 +39,7 @@ public class ReportController {
     )
     @GetMapping
     public String helloMessage(){
-        return HELLO_MESSAGE;
+        return HELLO_MESSAGE_OF_REPORT_CONTROLLER;
     }
 
     @Operation(
@@ -85,30 +69,6 @@ public class ReportController {
     }
 
     @Operation(
-            summary = "Find photo by reportId",
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "Found photo:",
-                            content = @Content(mediaType = MediaType.IMAGE_JPEG_VALUE)
-                    ),
-                    @ApiResponse(
-                            responseCode = "404",
-                            description = "If photo not found"
-                    )
-            },
-            tags = "Reports"
-    )
-    @GetMapping(value = "/{id}/photo")
-    public ResponseEntity<byte[]> findPhotoByReportId(@PathVariable Long id) {
-        PhotoOfPet photoOfPet = photoOfPetService.findPhotoByReportId(id);
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.parseMediaType(photoOfPet.getMediaType()));
-        headers.setContentLength(photoOfPet.getData().length);
-        return ResponseEntity.status(HttpStatus.OK).headers(headers).body(photoOfPet.getData());
-    }
-
-    @Operation(
             summary = "Add new report",
             requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(
                     description = "Add report",
@@ -130,37 +90,6 @@ public class ReportController {
     @PostMapping
     public Report addReport(@RequestBody Report report) {
         return reportService.addReport(report);
-    }
-
-    @Operation(
-            summary = "Add photo of pet to report",
-            requestBody = @io.swagger.v3.oas.annotations.parameters.RequestBody(
-                    description = "Add photo to report",
-                    content = @Content(mediaType = MediaType.MULTIPART_FORM_DATA_VALUE)
-            ),
-            responses = {
-                    @ApiResponse(
-                            responseCode = "200",
-                            description = "Add photo",
-                            content = @Content(
-                                    mediaType = MediaType.IMAGE_JPEG_VALUE
-                            )
-                    ),
-                    @ApiResponse(
-                            responseCode = "400",
-                            description = "Add photo is to big"
-                    )
-            },
-            tags = "Reports"
-    )
-    @PostMapping(value = "/{id}/photo", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    public ResponseEntity<String> upLoadPhotoOfPet(@PathVariable Long id, @RequestParam MultipartFile photo) throws IOException {
-        if (photo.getSize() > 1024 * 300) {
-            logger.warn("Warning: photo is to big");
-            return ResponseEntity.badRequest().body("File is to big");
-        }
-        photoOfPetService.uploadPhotoOfPet(id, photo);
-        return ResponseEntity.ok().build();
     }
 
     @Operation(
