+++ src/main/java/pro/sky/animal_shelter_telegram_bot/repository/PetOwnerRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pro.sky.animal_shelter_telegram_bot.model.PetOwner;
 
-import java.util.Collection;
 import java.util.Optional;
 
 @Repository
@@ -12,8 +11,5 @@ public interface PetOwnerRepository extends JpaRepository<PetOwner, Long> {
 
     void deleteById(Long id);
 
-    Collection<PetOwner> findPetOwnerById(Long id);
-
     Optional<PetOwner> findPetOwnerByChatId(Long chatId);
-
 }
