+++ src/main/java/pro/sky/animal_shelter_telegram_bot/repository/PhotoOfPetRepository.java
@@ -12,6 +12,4 @@ public interface PhotoOfPetRepository extends JpaRepository<PhotoOfPet, Long> {
     void deleteById(Long id);
 
     Optional<PhotoOfPet> findByReportId(Long reportId);
-
-
 }
