+++ src/main/java/pro/sky/animal_shelter_telegram_bot/repository/ReportRepository.java
@@ -4,7 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import pro.sky.animal_shelter_telegram_bot.model.Report;
 
-import java.time.LocalDate;
 import java.util.Optional;
 
 @Repository
